// frontend/src/app/model/predmet.model.ts
export class Predmet{
  naziv: String;
  sifra: String;
  tip: String;
  espb: String;
  smer: String;
  fond_casova: String;
  cilj: String;
  termin_predavanja: String;
  termin_vezbi: String;
  lab_vezbe: String;
  Propozicije: String;
  nastavnik_predavanje1: String;
  nastavnik_predavanje2: String;
  nastavnik_vezbe1: String;
  nastavnik_vezbe2: String;
  semestar: Number;

}